[#1087] Fixed some stylistic issues and split Panel component into bunch of smaller ones.
